Merge develop into master for v2.5.0 release

+++ externals/awsiot/aws-iot-device-sdk-embedded-C/include/aws_iot_log.h
@@ -67,7 +67,7 @@ extern "C" {
 	}
 #define FUNC_EXIT_RC(x)    \
 	{\
-	printf("FUNC_EXIT:   %s L#%d Return Code : %d \n", __func__, __LINE__, x);  \
+	printf("FUNC_EXIT:   %s L#%d Return Code : %d \n", __func__, __LINE__, (int)(x));  \
 	return x; \
 	}
 #else
